Split out types to avoid curcular imports. WiP on context.
